Removed feature flag. Added new process variable to support multiple warnings
