+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/handlers/WarningCaseEventHandler.java
@@ -1,16 +1,14 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.handlers;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.wacaseeventhandler.clients.LaunchDarklyFeatureFlagProvider;
 import uk.gov.hmcts.reform.wacaseeventhandler.clients.WorkflowApiClient;
-import uk.gov.hmcts.reform.wacaseeventhandler.config.features.FeatureFlag;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.CancellationActions;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.DmnValue;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.Warning;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.WarningValues;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.request.EvaluateDmnRequest;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.request.SendMessageRequest;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.response.CancellationEvaluateResponse;
@@ -34,22 +32,15 @@ import static uk.gov.hmcts.reform.wacaseeventhandler.domain.camunda.DmnValue.dmn
 
 @Service
 @Order(2)
-@Slf4j
 @SuppressWarnings({"PMD.DataflowAnomalyAnalysis", "unchecked"})
 public class WarningCaseEventHandler implements CaseEventHandler {
 
     private final AuthTokenGenerator serviceAuthGenerator;
     private final WorkflowApiClient workflowApiClient;
-    private final LaunchDarklyFeatureFlagProvider featureFlagProvider;
 
-    public WarningCaseEventHandler(
-        AuthTokenGenerator serviceAuthGenerator,
-        WorkflowApiClient workflowApiClient,
-        LaunchDarklyFeatureFlagProvider featureFlagProvider
-    ) {
+    public WarningCaseEventHandler(AuthTokenGenerator serviceAuthGenerator, WorkflowApiClient workflowApiClient) {
         this.serviceAuthGenerator = serviceAuthGenerator;
         this.workflowApiClient = workflowApiClient;
-        this.featureFlagProvider = featureFlagProvider;
     }
 
     @Override
@@ -82,72 +73,98 @@ public class WarningCaseEventHandler implements CaseEventHandler {
         Set<CancellationEvaluateResponse> emptyWarnings = new LinkedHashSet<>();
         Set<CancellationEvaluateResponse> ctgWarnings = new LinkedHashSet<>();
         Set<Warning> warnings = new LinkedHashSet<>();
-        final boolean isNonIacWarnings = featureFlagProvider.getBooleanValue(
-            FeatureFlag.WA_NON_IAC_WARNINGS
-            );
+
         results.stream()
             .filter(CancellationEvaluateResponse.class::isInstance)
             .map(CancellationEvaluateResponse.class::cast)
             .filter(result -> CancellationActions.WARN == CancellationActions.from(result.getAction().getValue()))
             .forEach(warnResponse -> {
-                if (isNonIacWarnings) {
-                    if (StringUtils.isEmpty(warnResponse.getWarningCode())
-                        || StringUtils.isEmpty(warnResponse.getWarningText() != null)) {
-                        emptyWarnings.add(warnResponse);
-                    } else {
-                        if (!StringUtils.isEmpty(warnResponse.getProcessCategories())
-                            || !StringUtils.isEmpty(warnResponse.getTaskCategories())) {
-                            final CancellationEvaluateResponse localWarnResponse =
-                                CancellationEvaluateResponse.builder()
-                                .processCategories(warnResponse.getProcessCategories())
-                                .taskCategories(warnResponse.getTaskCategories())
-                                .warningText(warnResponse.getWarningText())
-                                .warningCode(warnResponse.getWarningCode()).build();
-                            ctgWarnings.add(localWarnResponse);
-                        } else {
-                            final Warning warning = new Warning(
-                                warnResponse.getWarningCode().getValue(),
-                                warnResponse.getWarningText().getValue()
-                            );
-                            warnings.add(warning);
-                        }
-                    }
+                if (StringUtils.isEmpty(warnResponse.getWarningCode())
+                    || StringUtils.isEmpty(warnResponse.getWarningText() != null)) {
+                    emptyWarnings.add(warnResponse);
                 } else {
-                    DmnValue<String> taskCategories = warnResponse.getTaskCategories();
-                    DmnValue<String> processCategories = warnResponse.getProcessCategories();
-                    sendWarningMessage(
-                        eventInformation.getCaseId(),
-                        taskCategories,
-                        processCategories,
-                        warnResponse.getWarningCode(),
-                        warnResponse.getWarningText()
-                    );
+                    if (!StringUtils.isEmpty(warnResponse.getProcessCategories())
+                        || !StringUtils.isEmpty(warnResponse.getTaskCategories())) {
+                        final CancellationEvaluateResponse localWarnResponse = CancellationEvaluateResponse.builder()
+                            .processCategories(warnResponse.getProcessCategories())
+                            .taskCategories(warnResponse.getTaskCategories())
+                            .warningText(warnResponse.getWarningText())
+                            .warningCode(warnResponse.getWarningCode()).build();
+                        ctgWarnings.add(localWarnResponse);
+                    } else {
+                        final Warning warning = new Warning(
+                            warnResponse.getWarningCode().getValue(),
+                            warnResponse.getWarningText().getValue()
+                        );
+                        warnings.add(warning);
+                    }
                 }
             });
 
-        if (isNonIacWarnings) {
-            log.info("{} feature toggle enabled", FeatureFlag.WA_NON_IAC_WARNINGS);
-            NonIacWarningMessage warningMessage = new NonIacWarningMessage(
-                workflowApiClient, serviceAuthGenerator, eventInformation
+        processWarningResponse(emptyWarnings, warnings, ctgWarnings, eventInformation);
+    }
+
+    private void processWarningResponse(
+        Set<CancellationEvaluateResponse> emptyWarnings,
+        Set<Warning> warnings,
+        Set<CancellationEvaluateResponse> ctgWarnings,
+        EventInformation eventInformation
+
+    ) {
+        // scenario: event without warning attributes
+        if (!emptyWarnings.isEmpty()) {
+            emptyWarnings.forEach(response -> sendWarningMessage(
+                eventInformation.getCaseId(),
+                response.getTaskCategories(),
+                response.getProcessCategories(),
+                null)
             );
-            warningMessage.processWarningResponse(emptyWarnings, warnings, ctgWarnings);
+        }
+
+        // scenario: event without categories. Should contain unique warning attributes
+        if (!warnings.isEmpty()) {
+            WarningValues warningValues = new WarningValues();
+            warnings.stream().forEach(warning -> warningValues.getValues().add(warning));
+
+            sendWarningMessage(
+                eventInformation.getCaseId(),
+                null,
+                null,
+                warningValues.getValuesAsJson()
+            );
+        }
+        // scenario: event with categories and warning attributes
+        if (!ctgWarnings.isEmpty()) {
+            ctgWarnings.forEach(response -> {
+                final Warning warning = new Warning(response.getWarningCode().getValue(),
+                                                    response.getWarningText().getValue());
+
+                WarningValues warningValues = new WarningValues();
+                warningValues.getValues().add(warning);
+
+                sendWarningMessage(
+                    eventInformation.getCaseId(),
+                    response.getTaskCategories(),
+                    response.getProcessCategories(),
+                    warningValues.getValuesAsJson()
+                );
+            });
         }
     }
 
     private void sendWarningMessage(String caseReference,
                                     DmnValue<String> categories,
                                     DmnValue<String> processCategories,
-                                    DmnValue<String> warningCode,
-                                    DmnValue<String> warningText) {
+                                    String warningVariables) {
         Set<SendMessageRequest> warningMessageRequest =
             buildWarningMessageRequest(caseReference, categories,
-                                       processCategories, warningCode, warningText);
+                                       processCategories, warningVariables);
 
         warningMessageRequest.forEach(message -> {
-            if (message != null) {
-                workflowApiClient.sendMessage(serviceAuthGenerator.generate(), message);
+                if (message != null) {
+                    workflowApiClient.sendMessage(serviceAuthGenerator.generate(), message);
+                }
             }
-        }
 
         );
     }
@@ -156,16 +173,13 @@ public class WarningCaseEventHandler implements CaseEventHandler {
         String caseReference,
         DmnValue<String> categories,
         DmnValue<String> processCategories,
-        DmnValue<String> warningCode,
-        DmnValue<String> warningText
+        String warningVariables
     ) {
 
         SendMessageRequest oldFormatWarningMessage =
-            createOldFormatWarningMessage(caseReference, categories, processCategories,
-                                          warningCode, warningText);
+            createOldFormatWarningMessage(caseReference, categories, processCategories, warningVariables);
         SendMessageRequest warningMessage =
-            createWarningMessage(caseReference, categories, processCategories,
-                                 warningCode, warningText);
+            createWarningMessage(caseReference, categories, processCategories, warningVariables);
 
         return new HashSet<>(asList(oldFormatWarningMessage, warningMessage));
     }
@@ -181,8 +195,7 @@ public class WarningCaseEventHandler implements CaseEventHandler {
     private SendMessageRequest createWarningMessage(String caseReference,
                                                     DmnValue<String> categories,
                                                     DmnValue<String> processCategories,
-                                                    DmnValue<String> warningCode,
-                                                    DmnValue<String> warningText) {
+                                                    String warningVariables) {
 
         Map<String, DmnValue<?>> correlationKeys = new ConcurrentHashMap<>();
         correlationKeys.put("caseId", dmnStringValue(caseReference));
@@ -193,7 +206,7 @@ public class WarningCaseEventHandler implements CaseEventHandler {
                 .collect(Collectors.toList());
 
             categoriesToCancel.forEach(cat ->
-                                           correlationKeys.put("__processCategory__" + cat, dmnBooleanValue(true))
+                correlationKeys.put("__processCategory__" + cat, dmnBooleanValue(true))
             );
         }
 
@@ -203,12 +216,12 @@ public class WarningCaseEventHandler implements CaseEventHandler {
                 .collect(Collectors.toList());
 
             categoriesToCancel.forEach(cat ->
-                                           correlationKeys.put("__processCategory__" + cat, dmnBooleanValue(true))
+                correlationKeys.put("__processCategory__" + cat, dmnBooleanValue(true))
             );
         }
 
-        if (checkForWarningProperties(warningCode, warningText)) {
-            return addWarningsToProcessVariables(correlationKeys, warningCode, warningText);
+        if (!StringUtils.isEmpty(warningVariables)) {
+            return addWarningsToProcessVariables(correlationKeys, warningVariables);
         }
 
         return SendMessageRequest.builder()
@@ -231,8 +244,7 @@ public class WarningCaseEventHandler implements CaseEventHandler {
     private SendMessageRequest createOldFormatWarningMessage(String caseReference,
                                                              DmnValue<String> categories,
                                                              DmnValue<String> processCategories,
-                                                             DmnValue<String> warningCode,
-                                                             DmnValue<String> warningText) {
+                                                             String warningVariables) {
 
         if (processCategories == null) {
             Map<String, DmnValue<?>> correlationKeys = new ConcurrentHashMap<>();
@@ -242,8 +254,8 @@ public class WarningCaseEventHandler implements CaseEventHandler {
                 correlationKeys.put("taskCategory", categories);
             }
 
-            if (checkForWarningProperties(warningCode, warningText)) {
-                return addWarningsToProcessVariables(correlationKeys, warningCode, warningText);
+            if (!StringUtils.isEmpty(warningVariables)) {
+                return addWarningsToProcessVariables(correlationKeys, warningVariables);
             }
             return SendMessageRequest.builder()
                 .messageName(TASK_WARN.getMessageName())
@@ -270,25 +282,14 @@ public class WarningCaseEventHandler implements CaseEventHandler {
         return new EvaluateDmnRequest(variables);
     }
 
-    private boolean checkForWarningProperties(
-        DmnValue<String> warningCode, DmnValue<String> warningText) {
-        return warningCode != null && warningText != null;
-    }
-
     private SendMessageRequest addWarningsToProcessVariables(
         Map<String, DmnValue<?>> correlationKeys,
-        DmnValue<String> warningCode,
-        DmnValue<String> warningText) {
-
-        Map<String, DmnValue<?>> processVariables = Map.of(
-            "warningCode", warningCode,
-            "warningText", warningText
-        );
+        String warningVariables) {
 
         return SendMessageRequest.builder()
             .messageName(TASK_WARN.getMessageName())
             .correlationKeys(correlationKeys)
-            .processVariables(processVariables)
+            .processVariables(Map.of("warningsToAdd", new DmnValue<>(warningVariables, "String")))
             .all(true)
             .build();
     }
